+++ tabs/tab_4.py
@@ -1,8 +1,6 @@
-import dash
 import dash_core_components as dcc
 import dash_html_components as html
-from dash.dependencies import Input, Output
-import database as db
+from database import database as db
 
 options = []
