+++ tabs/tab_5.py
@@ -1,6 +1,6 @@
 import dash_core_components as dcc
 import dash_html_components as html
-import database as db
+from database import database as db
 
 candidates = db.select_database("SELECT Candidate.name as [name] "
                                 + "FROM Candidate"
